Restructure files and update CMakeLists.txt
