+++ tests/benchmark/utils/shared_info.h
@@ -23,9 +23,7 @@
 #include <string.h>
 #include <cstring>
 #include <string>
-
 #include <vector>
-
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <sys/ioctl.h>
